requirements.txt
pandas
pyarrow==11.0.0
datasets
weaviate-client==3.*